requirements.txt
# NAVADA - AI Jobs Intelligence Platform
# Project Dependencies

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0

# API and Web Requests
requests>=2.28.0
python-dotenv>=1.0.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Interactive plotting
plotly>=5.15.0

# Optional: Advanced analytics
scikit-learn>=1.3.0

# Optional: Data processing
openpyxl>=3.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0